+++ R/get_predictions_merMod.R
@@ -31,14 +31,14 @@ get_predictions_merMod <- function(model, fitfram, ci.lvl, linv, type, terms, ty
   } else {
 
 
-    fitfram$predicted <- stats::predict(
+    fitfram$predicted <- suppressWarnings(stats::predict(
       model,
       newdata = fitfram,
       type = "response",
       re.form = ref,
       allow.new.levels = TRUE,
       ...
-    )
+    ))
 
     if (se) {
       # get standard errors from variance-covariance matrix