+++ NEWS.md
@@ -1,3 +1,21 @@
+# ggeffects 0.10.0
+
+## General
+
+* Better support, including confidence intervals, for some of the already supported model types.
+
+## New supported models
+
+* `gamlss`, `geeglm` (package **geepack**), `lmrob` and `glmrob` (package **robustbase**), `rlmer` (package **robustlmm**), `rq` and `rqss` (package **quantreg**)
+
+## Changes to functions
+
+* The steps for specifying a range of values (e.g. `terms = "predictor [1:10]"`) can now be changed with `by`, e.g. `terms = "predictor [1:10 by=.5]"` (see also vignette _Marginal Effects at Specific Values_).
+
+## Bug fixes
+
+* Models with `AsIs`-conversion from division of two variables as dependent variable, e.g. `I(amount/frequency)`, now should work.
+
 # ggeffects 0.9.0
 
 ## General